renamed list commands
