+++ src/bin/esy.js
@@ -162,12 +162,12 @@ type Command = {
 
 const commandsByName: {[name: string]: () => Command} = {
   build: () => require('./esyBuild'),
-  'build-ls': () => require('./esyBuildLs'),
-  'list-libs': () => require('./esyListLibs'),
   release: () => require('./esyRelease'),
   config: () => require('./esyConfig'),
   install: () => require('./esyInstall'),
   add: () => require('./esyAdd'),
+  'ls-builds': () => require('./esyLsBuilds'),
+  'ls-libs': () => require('./esyLsLibs'),
   'init-global-sandbox': () => require('./esyInitGlobalSandbox'),
   'build-eject': () => require('./esyBuildEject'),
   'build-shell': () => require('./esyBuildShell'),
